fix(message)  : enables only my room !

+++ message.js
@@ -19,7 +19,10 @@ app.set("view engine", "ejs");
 
 
 const getChannel = (name) => {
-    return client.channels.cache.find(channel => channel.name === name)
+   
+    return client.channels.cache.find(channel => {
+        return channel.guild.id == "552497873116463107" && channel.name.startsWith(name)  
+    })
 }
 
 const getMessageType = (message) => {
@@ -47,33 +50,42 @@ client.on("ready", () => {
 
 
     client.on("message", (message) => {
-        const botCommand = ". ! - + /".split(' ')
-        if(message.author.bot ||  botCommand.some(e=> e===message.content[0])) return
-
-       
-        const channelName = message.channel.name
-        const messageType = getMessageType(message)
-        if(!messageType) return;
-        if(messageType != channelName) {
-            message.channel.send(`${message.author} มึงส่งผิด channel ละไอ่เด็กเหี้ย กูย้ายไป ${messageType} ให้แล้ว`).then( (message) => {
-                setTimeout( () => {
+        // enable only my discord 
+        if(message.guild.id == "552497873116463107") {
+            const botCommand = ". ! - + /".split(' ')
+            if(message.author.bot ||  botCommand.some(e=> e===message.content[0])) return
+    
+           
+            const channelName = message.channel.name
+            const messageType = getMessageType(message)
+            if(!messageType) return;
+            if(messageType != channelName) {
+                message.channel.send(`${message.author} มึงส่งผิด channel ละไอ่เด็กเหี้ย กูย้ายไป ${messageType} ให้แล้ว`).then( (message) => {
+                    setTimeout( () => {
+                        message.delete()
+                    }, 15 * 1000)
+                })
+                const sendTo = getChannel(messageType)
+                if(sendTo) {
+                    console.log('Not found target room.')
+                    if(messageType== "files") {
+                        for(let i of message.attachments){
+                            const file = new Discord.MessageAttachment(i[1].url)
+                            getChannel(messageType).send(file)
+                       
+                        }
+                        message.delete()
+                    }else{
+                    
+                        getChannel(messageType).send(  `${message.author}, ${message.content}`)
+                    }
                     message.delete()
-                }, 15 * 1000)
-            })
-            if(messageType== "files") {
-                for(let i of message.attachments){
-                    const file = new Discord.MessageAttachment(i[1].url)
-                    getChannel(messageType).send(file)
-               
                 }
-            }else{
-            
-                getChannel(messageType).send(  `${message.author}, ${message.content}`)
+                
+    
             }
-            message.delete()
-            
-
         }
+       
 		
 	});
